Front End Design
Need to add the on button click display vision page functionality and also need to update the website in which it suits the space team.
Finding a better quality GIF would be perfect for the banner background. Need to add animations as well.
This front end design includes the blog and single blog page.